Find common subdir between basedir and trace path
When this gem is packed into a war along with other gems and code (with
warbler), the gem is placed into a subdirectory matching how bundler
installed the gems. The current basedir computed by the gem doesn't
match anything, so all logs have the full path (which is really
long). So instead compute the longest common base directory the first
time anything is logged and use that.

This may not do the right thing in all cases, but I don't think it's any
worse than what's there right now.